testFixtures for user context
