+++ authz/build.gradle
@@ -14,4 +14,5 @@ dependencies {
 
     testImplementation("io.vertx:vertx-junit5")
     testImplementation("io.vertx:vertx-web-client")
+    testImplementation(testFixtures(project(':entities')))
 }
