+++ entities/build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id "com.flipkart.varadhi.java-library-conventions"
+    id 'java-test-fixtures'
 }
 dependencies {
     implementation(project(":common"))
