+++ server/build.gradle
@@ -51,6 +51,7 @@ dependencies {
 
     testImplementation(project(":pulsar"))
     testImplementation(testFixtures(project(":core")))
+    testImplementation(testFixtures(project(":entities")))
     testImplementation("org.apache.curator:curator-test")
     testImplementation("io.vertx:vertx-junit5")
     testImplementation("io.vertx:vertx-web-client")
